Clicks given labels in applyFilters()
Page changes when left hand filter bar checks are applied and subsequent checks do not get applied. Need to either reduce it to 1 clickable option or handle resuming the clicks on the new page. Also when the new page is created, extractMSCSearchResults seems to apply to the previous page, not the new one